ai/gen-ai-agents/whatsapp-ai-agent/files/code/requirements.txt
flask
python-dotenv
openai
aiohttp
requests